--- apptoolkit/src/main/java/com/d4rk/android/libs/apptoolkit/app/oboarding/domain/data/model/OnboardingThemeChoice.kt
@@ -1,7 +1,0 @@
-package com.d4rk.android.libs.apptoolkit.app.oboarding.domain.data.model
-
-import androidx.compose.ui.graphics.vector.ImageVector
-
-data class OnboardingThemeChoice(
-    val key : String , val displayName : String , val icon : ImageVector , val description : String
-)
