+++ apptoolkit/src/main/java/com/d4rk/android/libs/apptoolkit/app/onboarding/domain/data/model/ui/OnboardingPage.kt
@@ -1,0 +1,21 @@
+package com.d4rk.android.libs.apptoolkit.app.onboarding.domain.data.model.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.vector.ImageVector
+
+sealed class OnboardingPage {
+
+    data class DefaultPage(
+        val key: String,
+        val title: String,
+        val description: String,
+        val imageVector: ImageVector,
+        val isEnabled: Boolean = true
+    ) : OnboardingPage()
+
+    data class CustomPage(
+        val key: String,
+        val content: @Composable () -> Unit,
+        val isEnabled: Boolean = true
+    ) : OnboardingPage()
+}
