+++ apptoolkit/src/main/java/com/d4rk/android/libs/apptoolkit/app/onboarding/utils/helpers/CrashlyticsOnboardingStateManager.kt
@@ -1,0 +1,20 @@
+package com.d4rk.android.libs.apptoolkit.app.onboarding.utils.helpers
+
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+
+class CrashlyticsOnboardingStateManager {
+    companion object {
+        var showCrashlyticsDialog by mutableStateOf(true)
+            private set
+
+        fun openDialog() {
+            showCrashlyticsDialog = true
+        }
+
+        fun dismissDialog() {
+            showCrashlyticsDialog = false
+        }
+    }
+}
