+++ apptoolkit/src/main/java/com/d4rk/android/libs/apptoolkit/app/onboarding/utils/helpers/FinalOnboardingKonfettiState.kt
@@ -1,0 +1,9 @@
+package com.d4rk.android.libs.apptoolkit.app.onboarding.utils.helpers
+
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+
+object FinalOnboardingKonfettiState {
+    var hasKonfettiBeenShownGlobally by mutableStateOf(false)
+}
